Refactor Storage Program
