+++ programs/storage_api/src/storage_processor.rs
@@ -2,107 +2,12 @@
 //!  Receive mining proofs from miners, validate the answers
 //!  and give reward for good proofs.
 
-use crate::storage_contract::{CheckedProof, Proof, ProofStatus, StorageContract};
+use crate::storage_contract::StorageAccount;
 use crate::storage_instruction::StorageInstruction;
-use crate::{get_segment_from_entry, ENTRIES_PER_SEGMENT};
 use log::*;
-use serde::Serialize;
-use solana_sdk::account::{Account, KeyedAccount};
-use solana_sdk::hash::Hash;
+use solana_sdk::account::KeyedAccount;
 use solana_sdk::instruction::InstructionError;
-use solana_sdk::instruction::InstructionError::InvalidArgument;
 use solana_sdk::pubkey::Pubkey;
-use std::cmp;
-
-pub const TOTAL_VALIDATOR_REWARDS: u64 = 1;
-pub const TOTAL_REPLICATOR_REWARDS: u64 = 1;
-
-fn count_valid_proofs(proofs: &[CheckedProof]) -> u64 {
-    let mut num = 0;
-    for proof in proofs {
-        if let ProofStatus::Valid = proof.status {
-            num += 1;
-        }
-    }
-    num
-}
-
-/// Serialize account data
-fn store_contract<T>(account: &mut Account, contract: &T) -> Result<(), InstructionError>
-where
-    T: Serialize,
-{
-    if bincode::serialize_into(&mut account.data[..], contract).is_err() {
-        return Err(InstructionError::AccountDataTooSmall);
-    }
-    Ok(())
-}
-
-/// Deserialize account data
-fn read_contract(account: &Account) -> Result<StorageContract, InstructionError> {
-    if let Ok(storage_contract) = bincode::deserialize(&account.data) {
-        Ok(storage_contract)
-    } else {
-        Err(InstructionError::InvalidAccountData)
-    }
-}
-
-/// Deserialize account data but handle uninitialized accounts
-fn read_contract_with_default(
-    op: &StorageInstruction,
-    account: &Account,
-) -> Result<StorageContract, InstructionError> {
-    let mut storage_contract = read_contract(&account);
-    if let Ok(StorageContract::Default) = storage_contract {
-        match op {
-            StorageInstruction::SubmitMiningProof { .. } => {
-                storage_contract = Ok(StorageContract::ReplicatorStorage {
-                    proofs: vec![],
-                    reward_validations: vec![],
-                })
-            }
-            StorageInstruction::AdvertiseStorageRecentBlockhash { .. }
-            | StorageInstruction::ProofValidation { .. } => {
-                storage_contract = Ok(StorageContract::ValidatorStorage {
-                    entry_height: 0,
-                    hash: Hash::default(),
-                    lockout_validations: vec![],
-                    reward_validations: vec![],
-                })
-            }
-            StorageInstruction::ClaimStorageReward { .. } => Err(InvalidArgument)?,
-        }
-    }
-    storage_contract
-}
-
-/// Store the result of a proof validation into the replicator account
-fn store_validation_result(
-    account: &mut Account,
-    segment_index: usize,
-    status: ProofStatus,
-) -> Result<(), InstructionError> {
-    let mut storage_contract = read_contract(account)?;
-    match &mut storage_contract {
-        StorageContract::ReplicatorStorage {
-            proofs,
-            reward_validations,
-            ..
-        } => {
-            if segment_index > reward_validations.len() || reward_validations.is_empty() {
-                reward_validations.resize(cmp::max(1, segment_index), vec![]);
-            }
-            let result = proofs[segment_index].clone();
-            reward_validations[segment_index].push(CheckedProof {
-                proof: result,
-                status,
-            });
-        }
-        _ => return Err(InstructionError::InvalidAccountData),
-    }
-    store_contract(account, &storage_contract)?;
-    Ok(())
-}
 
 pub fn process_instruction(
     _program_id: &Pubkey,
@@ -115,207 +20,50 @@ pub fn process_instruction(
     // accounts_keys[0] must be signed
     if keyed_accounts[0].signer_key().is_none() {
         info!("account[0] is unsigned");
-        Err(InstructionError::GenericError)?;
+        Err(InstructionError::MissingRequiredSignature)?;
     }
 
-    if let Ok(syscall) = bincode::deserialize(data) {
-        let mut storage_contract =
-            read_contract_with_default(&syscall, &keyed_accounts[0].account)?;
-        match syscall {
-            StorageInstruction::SubmitMiningProof {
-                sha_state,
-                entry_height,
-                signature,
-            } => {
-                if let StorageContract::ReplicatorStorage { proofs, .. } = &mut storage_contract {
-                    if keyed_accounts.len() != 1 {
-                        // keyed_accounts[1] should be the main storage key
-                        // to access its data
-                        Err(InstructionError::InvalidArgument)?;
-                    }
+    let num_keyed_accounts = keyed_accounts.len();
+    let (me, rest) = keyed_accounts.split_at_mut(1);
+    let me = &mut me[0];
 
-                    let segment_index = get_segment_from_entry(entry_height);
-                    if segment_index > proofs.len() || proofs.is_empty() {
-                        proofs.resize(cmp::max(1, segment_index), Proof::default());
-                    }
-
-                    if segment_index > proofs.len() {
-                        // only possible if usize max < u64 max
-                        return Err(InstructionError::InvalidArgument);
-                    }
-
-                    debug!(
-                        "Mining proof submitted with contract {:?} entry_height: {}",
-                        sha_state, entry_height
-                    );
-
-                    let proof_info = Proof {
-                        id: *keyed_accounts[0].signer_key().unwrap(),
-                        sha_state,
-                        signature,
-                    };
-                    proofs[segment_index] = proof_info;
-                } else {
-                    Err(InstructionError::InvalidArgument)?;
-                }
+    match bincode::deserialize(data).map_err(|_| InstructionError::InvalidInstructionData)? {
+        StorageInstruction::SubmitMiningProof {
+            sha_state,
+            entry_height,
+            signature,
+        } => {
+            if num_keyed_accounts != 1 {
+                Err(InstructionError::InvalidArgument)?;
             }
-            StorageInstruction::AdvertiseStorageRecentBlockhash { hash, entry_height } => {
-                if let StorageContract::ValidatorStorage {
-                    entry_height: state_entry_height,
-                    hash: state_hash,
-                    reward_validations,
-                    lockout_validations,
-                } = &mut storage_contract
-                {
-                    if keyed_accounts.len() != 1 {
-                        // keyed_accounts[1] should be the main storage key
-                        // to access its data
-                        Err(InstructionError::InvalidArgument)?;
-                    }
-
-                    let original_segments = *state_entry_height / ENTRIES_PER_SEGMENT;
-                    let segments = entry_height / ENTRIES_PER_SEGMENT;
-                    debug!(
-                        "advertise new last id segments: {} orig: {}",
-                        segments, original_segments
-                    );
-                    if segments <= original_segments {
-                        return Err(InstructionError::InvalidArgument);
-                    }
-
-                    *state_entry_height = entry_height;
-                    *state_hash = hash;
-
-                    // move lockout_validations to reward_validations
-                    *reward_validations = lockout_validations.clone();
-                    lockout_validations.clear();
-                    lockout_validations.resize(segments as usize, Vec::new());
-                } else {
-                    return Err(InstructionError::InvalidArgument);
-                }
-            }
-            StorageInstruction::ClaimStorageReward { entry_height } => {
-                if keyed_accounts.len() != 1 {
-                    // keyed_accounts[1] should be the main storage key
-                    // to access its data
-                    Err(InstructionError::InvalidArgument)?;
-                }
-
-                if let StorageContract::ValidatorStorage {
-                    reward_validations, ..
-                } = &mut storage_contract
-                {
-                    let claims_index = get_segment_from_entry(entry_height);
-                    let _num_validations = count_valid_proofs(&reward_validations[claims_index]);
-                    // TODO can't just create lamports out of thin air
-                    // keyed_accounts[0].account.lamports += TOTAL_VALIDATOR_REWARDS * num_validations;
-                    reward_validations.clear();
-                } else if let StorageContract::ReplicatorStorage {
-                    reward_validations, ..
-                } = &mut storage_contract
-                {
-                    // if current tick height is a full segment away? then allow reward collection
-                    // storage needs to move to tick heights too, until then this makes little sense
-                    let current_index = get_segment_from_entry(tick_height);
-                    let claims_index = get_segment_from_entry(entry_height);
-                    if current_index <= claims_index || claims_index >= reward_validations.len() {
-                        debug!(
-                            "current {:?}, claim {:?}, rewards {:?}",
-                            current_index,
-                            claims_index,
-                            reward_validations.len()
-                        );
-                        return Err(InstructionError::InvalidArgument);
-                    }
-                    let _num_validations = count_valid_proofs(&reward_validations[claims_index]);
-                    // TODO can't just create lamports out of thin air
-                    // keyed_accounts[0].account.lamports += num_validations
-                    //     * TOTAL_REPLICATOR_REWARDS
-                    //     * (num_validations / reward_validations[claims_index].len() as u64);
-                    reward_validations.clear();
-                } else {
-                    return Err(InstructionError::InvalidArgument);
-                }
-            }
-            StorageInstruction::ProofValidation {
-                entry_height: proof_entry_height,
-                proofs,
-            } => {
-                if let StorageContract::ValidatorStorage {
-                    entry_height: current_entry_height,
-                    lockout_validations,
-                    ..
-                } = &mut storage_contract
-                {
-                    if keyed_accounts.len() == 1 {
-                        // have to have at least 1 replicator to do any verification
-                        Err(InstructionError::InvalidArgument)?;
-                    }
-
-                    if proof_entry_height >= *current_entry_height {
-                        return Err(InstructionError::InvalidArgument);
-                    }
-
-                    let segment_index = get_segment_from_entry(proof_entry_height);
-                    let mut previous_proofs = keyed_accounts[1..]
-                        .iter_mut()
-                        .filter_map(|account| {
-                            read_contract(&account.account)
-                                .ok()
-                                .map(move |contract| match contract {
-                                    StorageContract::ReplicatorStorage { proofs, .. } => {
-                                        Some((&mut account.account, proofs[segment_index].clone()))
-                                    }
-                                    _ => None,
-                                })
-                        })
-                        .flatten()
-                        .collect::<Vec<_>>();
-
-                    if previous_proofs.len() != proofs.len() {
-                        // don't have all the accounts to validate the proofs against
-                        return Err(InstructionError::InvalidArgument);
-                    }
-
-                    let mut valid_proofs: Vec<_> = proofs
-                        .into_iter()
-                        .enumerate()
-                        .filter_map(|(i, entry)| {
-                            if previous_proofs[i].1.signature != entry.proof.signature
-                                || entry.status != ProofStatus::Valid
-                            {
-                                let _ = store_validation_result(
-                                    &mut previous_proofs[i].0,
-                                    segment_index,
-                                    entry.status,
-                                );
-                                None
-                            } else if store_validation_result(
-                                &mut previous_proofs[i].0,
-                                segment_index,
-                                entry.status.clone(),
-                            )
-                            .is_ok()
-                            {
-                                Some(entry)
-                            } else {
-                                None
-                            }
-                        })
-                        .collect();
-
-                    // allow validators to store successful validations
-                    lockout_validations[segment_index].append(&mut valid_proofs);
-                } else {
-                    return Err(InstructionError::InvalidArgument);
-                }
-            }
+            me.submit_mining_proof(sha_state, entry_height, signature)
         }
-        store_contract(&mut keyed_accounts[0].account, &storage_contract)?;
-        Ok(())
-    } else {
-        info!("Invalid instruction data: {:?}", data);
-        Err(InstructionError::InvalidInstructionData)
+        StorageInstruction::AdvertiseStorageRecentBlockhash { hash, entry_height } => {
+            if num_keyed_accounts != 1 {
+                // keyed_accounts[0] should be the main storage key
+                // to access its data
+                Err(InstructionError::InvalidArgument)?;
+            }
+            me.advertise_storage_recent_blockhash(hash, entry_height)
+        }
+        StorageInstruction::ClaimStorageReward { entry_height } => {
+            if num_keyed_accounts != 1 {
+                // keyed_accounts[0] should be the main storage key
+                // to access its data
+                Err(InstructionError::InvalidArgument)?;
+            }
+            me.claim_storage_reward(entry_height, tick_height)
+        }
+        StorageInstruction::ProofValidation {
+            entry_height,
+            proofs,
+        } => {
+            if num_keyed_accounts == 1 {
+                // have to have at least 1 replicator to do any verification
+                Err(InstructionError::InvalidArgument)?;
+            }
+            me.proof_validation(entry_height, proofs, rest)
+        }
     }
 }
 
@@ -323,6 +71,7 @@ pub fn process_instruction(
 mod tests {
     use super::*;
     use crate::id;
+    use crate::storage_contract::{CheckedProof, Proof, ProofStatus, State, StorageContract};
     use crate::storage_instruction;
     use crate::ENTRIES_PER_SEGMENT;
     use bincode::deserialize;
@@ -430,7 +179,7 @@ mod tests {
         let pubkey = &Pubkey::default();
         let mut keyed_account = KeyedAccount::new(&pubkey, false, &mut account);
         // pretend it's a validator op code
-        let mut contract = read_contract(&keyed_account.account).unwrap();
+        let mut contract = keyed_account.state().unwrap();
         if let StorageContract::ValidatorStorage { .. } = contract {
             assert!(true)
         }
@@ -444,7 +193,7 @@ mod tests {
             lockout_validations: vec![],
             reward_validations: vec![],
         };
-        store_contract(&mut keyed_account.account, &contract).unwrap();
+        keyed_account.set_state(&contract).unwrap();
         if let StorageContract::ReplicatorStorage { .. } = contract {
             panic!("this shouldn't work");
         }
@@ -452,7 +201,7 @@ mod tests {
             proofs: vec![],
             reward_validations: vec![],
         };
-        store_contract(&mut keyed_account.account, &contract).unwrap();
+        keyed_account.set_state(&contract).unwrap();
         if let StorageContract::ValidatorStorage { .. } = contract {
             panic!("this shouldn't work");
         }
@@ -658,33 +407,5 @@ mod tests {
             get_storage_blockhash(&bank, &validator_pubkey),
             storage_blockhash
         );
-    }
-
-    /// check that uninitialized accounts are handled
-    #[test]
-    fn test_read_contract_with_default() {
-        let mut account = Account::default();
-        // no space allocated
-        assert!(read_contract(&account).is_err());
-        account.data.resize(4 * 1024, 0);
-        let instruction = StorageInstruction::AdvertiseStorageRecentBlockhash {
-            hash: Hash::default(),
-            entry_height: 0,
-        };
-        read_contract_with_default(&instruction, &account).unwrap();
-        let instruction = StorageInstruction::SubmitMiningProof {
-            sha_state: Hash::default(),
-            entry_height: 0,
-            signature: Signature::default(),
-        };
-        read_contract_with_default(&instruction, &account).unwrap();
-        let instruction = StorageInstruction::ProofValidation {
-            entry_height: 0,
-            proofs: vec![],
-        };
-        read_contract_with_default(&instruction, &account).unwrap();
-        // Can't claim rewards on an uninitialized account
-        let instruction = StorageInstruction::ClaimStorageReward { entry_height: 0 };
-        assert!(read_contract_with_default(&instruction, &account).is_err());
     }
 }
